# backend/src/schemas.py
"""Pydantic request/response schemas."""

from pydantic import BaseModel, Field, EmailStr
from datetime import datetime
from typing import Optional


# ============ Authentication Schemas ============


class UserSignup(BaseModel):
    """Schema for user signup request."""

    email: EmailStr = Field(..., description="User email address")
    password: str = Field(..., min_length=8, description="Password (min 8 chars)")

    class Config:
        json_schema_extra = {
            "example": {
                "email": "user@example.com",
                "password": "SecurePass123!"
            }
        }


class UserLogin(BaseModel):
    """Schema for user login request."""

    email: EmailStr = Field(..., description="User email address")
    password: str = Field(..., description="User password")

    class Config:
        json_schema_extra = {
            "example": {
                "email": "user@example.com",
                "password": "SecurePass123!"
            }
        }


class AuthResponse(BaseModel):
    """Schema for authentication response (signup/login)."""

    access_token: str = Field(..., description="JWT access token")
    token_type: str = Field(default="bearer", description="Token type")
    user_id: str = Field(..., description="User ID (from JWT sub claim)")

    class Config:
        json_schema_extra = {
            "example": {
                "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                "token_type": "bearer",
                "user_id": "user-123"
            }
        }


class VerifyResponse(BaseModel):
    """Schema for token verification response."""

    valid: bool = Field(..., description="Whether token is valid")
    user_id: Optional[str] = Field(None, description="User ID if token is valid")

    class Config:
        json_schema_extra = {
            "example": {
                "valid": True,
                "user_id": "user-123"
            }
        }


# ============ Task Schemas ============


class TaskCreate(BaseModel):
    """Schema for creating a new task."""

    title: str = Field(min_length=1, max_length=200)
    description: Optional[str] = None


class TaskUpdate(BaseModel):
    """Schema for updating an existing task."""

    title: Optional[str] = Field(None, min_length=1, max_length=200)
    description: Optional[str] = None


class TaskResponse(BaseModel):
    """Schema for task response (full task object)."""

    id: int
    user_id: str
    title: str
    description: Optional[str]
    completed: bool
    created_at: datetime
    updated_at: datetime

    class Config:
        """Pydantic config for ORM compatibility."""
        from_attributes = True